Extracted files, cr fixes
